Fixups for the Windows build
It turns out that we need the -MD flag, although the symptom resulting
from its lack (msgfmt.exe crashing with STACK_BUFFER_OVERRUN error) was
strange to me.

Also remove the dependency on libffi. This has been in the recipe from
the very start, but as far as I can tell has never been used.