Unsuccessful efforts to fix the Windows build
Here are some updates to the build infrastructure to get 0.23.1 building
on Windows with MSVC. However, my takeaway (see #35, as well as #34, #36,
and #37) is that some of the updates in 0.23.1 may just fundamentally break
the ability to build gettext on this platform. I *was* able to get it to
build on MinGW/GCC, but that would be a major ABI compatibility change that
we don't want to undertake right now.